requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic-settings==2.1.0
yfinance==0.2.33
pandas==2.1.4
pytest==7.4.3
lxml==6.0.2
google-cloud-bigquery
google-genai
google-adk
db-dtypes